feat: update existing file docs between tags

+++ src/lib/index.ts
@@ -1,4 +1,4 @@
-import {writeFileSync} from 'fs';
+import {writeFileSync, existsSync, readFileSync} from 'fs';
 import {buildDocumentation} from './docs';
 import {documentationToMarkdown} from './markdown';
 import type {DocEntry, DocEntryConstructor, DocEntryType} from './types';
@@ -26,6 +26,16 @@ export const generateDocumentation = ({
   });
 
   const markdown: string = documentationToMarkdown(entries);
+
+  if (existsSync(outputFile)) {
+    const fileContent = readFileSync(outputFile, 'utf-8');
+
+    const regex = /<!-- TSDOC_START -->([^]*?)<!-- TSDOC_END -->/gm;
+    const replace = `<!-- TSDOC_START -->\n\n${markdown}\n<!-- TSDOC_END -->`
+    writeFileSync(outputFile, fileContent.replace(regex, replace), 'utf-8');
+
+    return;
+  }
 
   writeFileSync(outputFile, markdown, 'utf-8');
 };
